refactor(open): move workspace enable and sync logic to services
- Replace direct config file parsing with ConfigManager usage
- Move workspace enable logic into ConfigManager.enableWorkspace method
- Move workspace sync logic into WorkspaceManager.checkoutWorkspace method
- Remove deprecated enableWorkspace and syncSingleWorkspace functions
- Add detailed error handling and success logging in openCommand
- Improve code structure by introducing ConfigManager and WorkspaceManager classes
- Ensure config is written after enabling workspace in config manager
