+++ src/services/workspace-manager.ts
